backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
PyMuPDF==1.23.8
pandas==2.1.4
python-docx==0.8.11
openpyxl==3.1.2
python-dotenv==1.0.0
asyncio-mqtt==0.13.0
requests==2.31.0
langchain==0.1.16
langchain-community==0.0.32
# Added for embedding support
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
tiktoken==0.5.2 